requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.txt
#
cairocffi==0.7.2
cairosvg==1.0.22
certifi==2018.1.18        # via requests
cffi==1.8.3
chardet==3.0.4            # via requests
click==6.6
cssselect==0.9.2
#django==2.0.3
Django==1.11.11
html5lib==0.999999999
idna==2.6                 # via requests
itsdangerous==0.24
lxml==3.6.4
markupsafe==0.23
pycparser==2.14
pyphen==0.9.4
python-dateutil==2.6.0
python-quickbooks==0.5.7
pytz==2018.3
rauth==0.7.3              # via python-quickbooks
redis==2.10.5
requests==2.18.4          # via python-quickbooks, rauth
simplejson==3.13.2        # via python-quickbooks
six==1.10.0
tinycss==0.3
urllib3==1.22             # via requests
weasyprint==0.31
webencodings==0.5
werkzeug==0.11.11
dj-database-url==0.4.2
psycopg2==2.7.4
bpython==0.17.1
django-filter==1.1.0
djangorestframework==3.7.7
